@@ -8,7 +8,7 @@ import io.flutter.plugin.common.PluginRegistry.Registrar
 import java.net.NetworkInterface
 import java.util.*
 
-class GetIpPlugin(): MethodCallHandler {
+class GetIpPlugin : MethodCallHandler {
   companion object {
     @JvmStatic
     fun registerWith(registrar: Registrar): Unit {
@@ -27,14 +27,14 @@ class GetIpPlugin(): MethodCallHandler {
     }
   }
 
-  fun getIPAddress(useIPv4: Boolean): String {
+  private fun getIPAddress(useIPv4: Boolean): String {
     try {
       val interfaces = Collections.list(NetworkInterface.getNetworkInterfaces())
       for (intf in interfaces) {
-        val addrs = Collections.list(intf.getInetAddresses())
+        val addrs = Collections.list(intf.inetAddresses)
         for (addr in addrs) {
-          if (!addr.isLoopbackAddress()) {
-            val sAddr = addr.getHostAddress()
+          if (!addr.isLoopbackAddress) {
+            val sAddr = addr.hostAddress
             //boolean isIPv4 = InetAddressUtils.isIPv4Address(sAddr);
             val isIPv4 = sAddr.indexOf(':') < 0
 
@@ -44,14 +44,14 @@ class GetIpPlugin(): MethodCallHandler {
             } else {
               if (!isIPv4) {
                 val delim = sAddr.indexOf('%') // drop ip6 zone suffix
-                return if (delim < 0) sAddr.toUpperCase() else sAddr.substring(0, delim).toUpperCase()
+                return if (delim < 0) sAddr.uppercase() else sAddr.substring(0, delim).uppercase()
               }
             }
           }
         }
       }
     } catch (e: Exception) {
-      print(e);
+      print(e)
     }
     return ""
   }
